LPS-140253 Rename to LayoutVisibilityManager
